feat: 답변(사용자인증로직 추가) 구현중 [be01]

+++ server/src/main/java/com/preproject/stackOverflow/question/dto/QuestionDto.java
@@ -64,7 +64,7 @@ public class QuestionDto {
         private String content;
         private String tag;
         private List<String> tags;
-        private int vote;
+        private long vote;
         private LocalDateTime createdAt;
         private LocalDateTime modifiedAt;
         private Question.QuestionStatus questionStatus;
